pyproject.toml
[project]
name = "trainer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "diffusers[torch]",
    "hydra-core>=1.3.2",
    "mosaicml-streaming>=0.13.0",
    "peft>=0.18.1",
    "ray[train]>=2.54.0",
    "tensorboard>=2.20.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers<5",
]

[project.scripts]
trainer = "trainer.main:main"

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"

[tool.uv.sources]
diffusers = { git = "https://github.com/huggingface/diffusers.git" }
